[frontend] clean
